hub: split server bootstrap out of hub.rs
No behavior change.
